Move generateSchema into the SchemaGenerator
