+++ graphql-kotlin-schema-generator/src/main/kotlin/com/expediagroup/graphql/generator/SchemaGenerator.kt
@@ -17,18 +17,29 @@
 package com.expediagroup.graphql.generator
 
 import com.expediagroup.graphql.SchemaGeneratorConfig
+import com.expediagroup.graphql.TopLevelObject
 import com.expediagroup.graphql.directives.DeprecatedDirective
 import com.expediagroup.graphql.generator.state.ClassScanner
 import com.expediagroup.graphql.generator.state.TypesCache
 import com.expediagroup.graphql.generator.types.generateGraphQLType
+import com.expediagroup.graphql.generator.types.generateMutations
+import com.expediagroup.graphql.generator.types.generateQueries
+import com.expediagroup.graphql.generator.types.generateSubscriptions
 import graphql.Directives
 import graphql.schema.GraphQLCodeRegistry
 import graphql.schema.GraphQLDirective
+import graphql.schema.GraphQLSchema
 import graphql.schema.GraphQLType
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.reflect.KClass
 import kotlin.reflect.full.createType
 
+/**
+ * Generate a schema object given some configuration and top level objects for the queries, mutaitons, and subscriptions.
+ *
+ * This class maintains the state of the schema while generation is taking place. It is passed into the internal functions
+ * so they can use the cache and add additional types and directives into the schema as they parse the Kotlin code.
+ */
 open class SchemaGenerator(internal val config: SchemaGeneratorConfig) {
 
     internal val classScanner = ClassScanner(config.supportedPackages)
@@ -47,6 +58,30 @@ open class SchemaGenerator(internal val config: SchemaGeneratorConfig) {
         // @deprecated directive is a built-in directive that each GraphQL server should provide bu currently it is not added by graphql-java
         //   see https://github.com/graphql-java/graphql-java/issues/1598
         directives[DeprecatedDirective.name] = DeprecatedDirective
+    }
+
+    /**
+     * Generate a schema given a list of objects to parse for the queries, mutations, and subscriptions.
+     */
+    fun generateSchema(
+        queries: List<TopLevelObject>,
+        mutations: List<TopLevelObject> = emptyList(),
+        subscriptions: List<TopLevelObject> = emptyList()
+    ): GraphQLSchema {
+        val builder = GraphQLSchema.newSchema()
+        builder.query(generateQueries(this, queries))
+        builder.mutation(generateMutations(this, mutations))
+        builder.subscription(generateSubscriptions(this, subscriptions))
+
+        // add unreferenced interface implementations
+        additionalTypes.forEach {
+            builder.additionalType(it)
+        }
+
+        builder.additionalDirectives(directives.values.toSet())
+        builder.codeRegistry(codeRegistry.build())
+
+        return config.hooks.willBuildSchema(builder).build()
     }
 
     /**
